Fix crash on iPad
Without presentation controller actionSheet fails on iPad with uncaught exception:

[ Uncaught Exception ]
Name: NSGenericException, Reason: Your application has presented a UIAlertController (<UIAlertController:>) of style UIAlertControllerStyleActionSheet. The modalPresentationStyle of a UIAlertController with this style is UIModalPresentationPopover. You must provide location information for this popover through the alert controller's popoverPresentationController. You must provide either a sourceView and sourceRect or a barButtonItem.  If this information is not known when you present the alert controller, you may provide it in the UIPopoverPresentationControllerDelegate method -prepareForPopoverPresentation.